fixup! fixup! fixup! remove redundant methods
