+++ src/snowcli/cli/snowpark/common.py
@@ -1,0 +1,36 @@
+import sys
+from typing import TextIO
+
+
+if not sys.stdout.closed and sys.stdout.isatty():
+    GREEN = "\033[32m"
+    BLUE = "\033[34m"
+    ORANGE = "\033[38:2:238:76:44m"
+    GRAY = "\033[2m"
+    ENDC = "\033[0m"
+else:
+    GREEN = ""
+    ORANGE = ""
+    BLUE = ""
+    GRAY = ""
+    ENDC = ""
+
+
+def _prefix_line(prefix: str, line: str) -> str:
+    """
+    _prefix_line ensure the prefix is still present even when dealing with return characters
+    """
+    if "\r" in line:
+        line = line.replace("\r", f"\r{prefix}")
+    if "\n" in line[:-1]:
+        line = line[:-1].replace("\n", f"\n{prefix}") + line[-1:]
+    if not line.startswith("\r"):
+        line = f"{prefix}{line}"
+    return line
+
+
+def print_log_lines(file: TextIO, name, id, logs):
+    prefix = f"{GREEN}{name}/{id}{ENDC} "
+    logs = logs[0:-1]
+    for log in logs:
+        print(_prefix_line(prefix, log + "\n"), file=file, end="", flush=True)
